feat(schema): move relations to separate files
