+++ src/db/schema/donors.ts
@@ -1,4 +1,3 @@
-import { relations } from 'drizzle-orm';
 import {
   integer,
   pgTable,
@@ -7,8 +6,8 @@ import {
   text,
 } from 'drizzle-orm/pg-core';
 import { createdAt, updatedAt } from './shared';
+import { needs } from './needs';
 import { users } from './users';
-import { needs } from '@/db/schema/needs';
 
 export const donors = pgTable('donors', {
   id: serial('id').primaryKey(),
@@ -37,25 +36,4 @@ export const donorFavouriteNeeds = pgTable(
       name: 'donor_favourite_needs_pkey',
     }),
   ],
-);
-
-export const donorsRelations = relations(donors, ({ one }) => ({
-  user: one(users, {
-    fields: [donors.userId],
-    references: [users.id],
-  }),
-}));
-
-export const donorFavouriteNeedsRelations = relations(
-  donorFavouriteNeeds,
-  ({ one }) => ({
-    donor: one(donors, {
-      fields: [donorFavouriteNeeds.donorId],
-      references: [donors.id],
-    }),
-    need: one(needs, {
-      fields: [donorFavouriteNeeds.needId],
-      references: [needs.id],
-    }),
-  }),
 );
