+++ src/db/schema/index.ts
@@ -1,6 +1,13 @@
 export * from './shared';
+export * from './addresses';
 export * from './users';
+export * from './users.relations';
 export * from './donors';
+export * from './donors.relations';
 export * from './organizations';
+export * from './addresses.relations';
+export * from './organizations.relations';
 export * from './needs';
+export * from './needs.relations';
 export * from './donations';
+export * from './donations.relations';
